+++ packages/core/navigation-next/src/components/LayoutManagerWithViewController/types.js
@@ -5,9 +5,11 @@ import type { WithAnalyticsEventsProps } from '@atlaskit/analytics-next';
 
 import { UIController, ViewController } from '../../';
 
+import type { CollapseListeners } from '../LayoutManager/types';
+
 import type { ViewControllerState } from '../../view-controller/types';
 
-export type LayoutManagerWithViewControllerProps = {
+export type LayoutManagerWithViewControllerProps = CollapseListeners & {
   children: Node,
   customComponents: { [string]: ComponentType<*> },
   globalNavigation: ComponentType<{}>,